% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_metabolism_estimates.R
\name{download_metabolism_estimates}
\alias{download_metabolism_estimates}
\title{Download metabolism estimates and predictors}
\usage{
download_metabolism_estimates(save_dir = temp, overwrite_file = FALSE)
}
\arguments{
\item{save_dir}{Path to save .tsv file containing daily metabolism estimates. If save_dir isn't specified, metabolism estimates will not be saved locally.}

\item{overwrite_file}{Boolean indicating if daily_predictions.tsv file should be overwritten if it already exists. Default is FALSE.}
}
\value{
Returns one data frame containing daily metabolism estimates for 356 rivers.
}
\description{
Download daily metabolism estimates and potential predictor variables for 356 rivers. See metadata associated with data release item 8: \href{https://www.sciencebase.gov/catalog/item/59eb9c0ae4b0026a55ffe389}{Metabolism estimates and predictors}.
}
\examples{
download_metabolism_estimates()

\dontrun{
download_metabolism_estimates(save_dir = "./data/daily_estimates",overwrite_file=TRUE)
}

}
\keyword{estimates}
\keyword{predictor}
