% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_model_config.R
\name{download_model_config}
\alias{download_model_config}
\title{Download metabolism model configurations}
\usage{
download_model_config()
}
\value{
Returns one data frame with one row per model.
}
\description{
Download model specifications used for estimating metabolism. This function will download and load a table containing one row per model. See metadata associated with data release item 5: \href{https://www.sciencebase.gov/catalog/item/59eb9b86e4b0026a55ffe379}{Model configurations}.
}
\examples{
download_model_config()


}
