% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_model_diagnostics.R
\name{download_model_diagnostics}
\alias{download_model_diagnostics}
\title{Download model diagnostics}
\usage{
download_model_diagnostics()
}
\value{
Returns a data frame with one row per model.
}
\description{
Download diagnostics and assessments of metabolism model performance. See metadata associated with data release item 7: \href{https://www.sciencebase.gov/catalog/item/59eb9bafe4b0026a55ffe382}{Model diagnostics}.
}
\examples{
download_model_diagnostics()


}
