% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_site_data.R
\name{download_site_data}
\alias{download_site_data}
\title{Download site data}
\usage{
download_site_data(save_dir = temp)
}
\arguments{
\item{save_dir}{Path to save .tsv file containing site data. If save_dir isn't specified, site data will not be saved locally.}
}
\value{
Returns a data frame containing one row of site identifiers, details, and quality indicators per site
}
\description{
Download site information for all sites contained within "Metabolic estimates for 356 rivers (2007-2017), \url{https://doi.org/10.5066/F70864KX}." See metadata associated with data item 1: \href{https://www.sciencebase.gov/catalog/item/59bff64be4b091459a5e098b}{Site data}.
}
\examples{
download_site_data()

\dontrun{

pc_sites <- download_site_data(save_dir = "./data/site_data")
}

}
\keyword{site}
