% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_timeseries_data.R
\name{download_timeseries_data}
\alias{download_timeseries_data}
\title{Download time series data}
\usage{
download_timeseries_data(
sitename,
save_dir,
overwrite_file = FALSE
)
}
\arguments{
\item{sitename}{The sitename should be in the same format used in the site data. See examples below.}

\item{save_dir}{Path of the desired save directory}

\item{overwrite_file}{Boolean indicating if local file should be overwritten if it already exists. (\code{\link[sbtools]{item_file_download}})}
}
\value{
Downloads one zip file per site to save_dir.
}
\description{
Download time series data for specified sites. Time series data contains data on water quality and quantity, collected or computed from outside sources. This function will download 1 zip file per site to the specified save directory. See metadata associated with data release item 3: \href{https://www.sciencebase.gov/catalog/item/59c03b72e4b091459a5e0b64}{timeseries data}.
}
\examples{
\dontrun{

download_timeseries_data("nwis_01036390","./data/timeseries")
download_timeseries_data(sitename = c("nwis_01036390","nwis_01570500"),
                          save_dir = "./data/timeseries",TRUE)

pc_sites <- download_site_data()
lapply(pc_sites$site_name,download_timeseries_data,
       save_dir = "./data/timeseries",overwrite_file=TRUE)

}

}
\keyword{series}
\keyword{time}
\keyword{timeseries}
