% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_url.R
\name{get_spatial_url}
\alias{get_spatial_url}
\title{Get spatial url}
\usage{
get_spatial_url()
}
\value{
Returns a character string containing the web url to access spatial data
}
\description{
While direct download of spatial data (data release \href{https://www.sciencebase.gov/catalog/item/5a5e3791e4b06e28e9be48c1}{item 2} containing spatial coordinates and site catchment boundaries) is not currently enabled, this function returns the web url where spatial data can be downloaded manually.
}
\examples{
get_spatial_url
}
\keyword{spatial}
