% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_timeseries_data.R
\name{load_timeseries_data}
\alias{load_timeseries_data}
\title{Load time series data}
\usage{
load_timeseries_data(
sitename,
zip_dir
)
}
\arguments{
\item{sitename}{The sitename should be in the same format used in the site data. See examples below.}

\item{zip_dir}{Path indicating location of downloaded timeseries.zip files. See download_time_series_data().}
}
\value{
Returns a list for one site containing data frames housing the different water quality and quantity data sets.
}
\description{
Load time series data for a specified site. Timeseries data contains data on water quality and quantity, collected or computed from outside sources. See metadata associated with data release item 3: \href{https://www.sciencebase.gov/catalog/item/59c03b72e4b091459a5e0b64}{timeseries data}.
}
\examples{
\dontrun{

# Timeseries zip files for selected sites must have been previously downloaded,
# see download_timeseries_data():
load_timeseries_data("nwis_01036390",zip_dir = "./data/timeseries")

# Load timeseries data from multiple sites in a nested list:
selected_sites <- c("nwis_01021050","nwis_01125100")
selected_sites_ls <- lapply(selected_sites,load_timeseries_data,zip_dir = "./data/timeseries")
names(selected_sites_ls) <- selected_sites
}
}
