% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_model_outputs.R
\name{download_model_outputs}
\alias{download_model_outputs}
\title{Download metabolism model outputs}
\usage{
download_model_outputs(sitename, save_dir = temp, overwrite_file = FALSE)
}
\arguments{
\item{sitename}{The sitename should be in the same format used in the site data. See examples below.}

\item{save_dir}{Path to save .zip file containing model outputs. If save_dir isn't specified, the zip file containing model outputs will not be saved locally.}

\item{overwrite_file}{Boolean indicating if .zip folder should be overwritten if it already exists. Default is FALSE.}
}
\value{
Returns a list containing five data frames per site that describe the model fits.
}
\description{
Download complete fits from metabolism estimation models. This function will download and load five data frames per model that collectively describe the model fits. See ScienceBase for a complete description of model output files. See metadata associated with data release item 6: \href{https://www.sciencebase.gov/catalog/item/59eb9ba6e4b0026a55ffe37f}{Model outputs}.
}
\examples{
download_model_outputs("nwis_01184000")


\dontrun{

download_model_outputs("nwis_01184000",save_dir = "./data/model_outputs")
}


}
\keyword{model}
\keyword{output}
